[FLINK-20455] Add check for jar file contents
