+++ tools/ci/compile.sh
@@ -29,6 +29,10 @@ fi
 CI_DIR="$HERE/../ci"
 MVN_CLEAN_COMPILE_OUT="/tmp/clean_compile.out"
 
+# Deploy into this directory, to run license checks on all jars staged for deployment.
+# This helps us ensure that ALL artifacts we deploy to maven central adhere to our license conditions.
+MVN_VALIDATION_DIR="/tmp/flink-validation-deployment"
+
 # source required ci scripts
 source "${CI_DIR}/stage.sh"
 source "${CI_DIR}/shade.sh"
@@ -43,7 +47,7 @@ echo "==========================================================================
 
 EXIT_CODE=0
 
-run_mvn clean install $MAVEN_OPTS -Dflink.convergence.phase=install -Pcheck-convergence -Dflink.forkCount=2 \
+run_mvn clean deploy -DaltDeploymentRepository=validation_repository::default::file:$MVN_VALIDATION_DIR $MAVEN_OPTS -Dflink.convergence.phase=install -Pcheck-convergence -Dflink.forkCount=2 \
     -Dflink.forkCountTestPackage=2 -Dmaven.javadoc.skip=true -U -DskipTests | tee $MVN_CLEAN_COMPILE_OUT
 
 EXIT_CODE=${PIPESTATUS[0]}
@@ -107,8 +111,10 @@ EXIT_CODE=$(($EXIT_CODE+$?))
 
 echo "============ Run license check ============"
 
+find $MVN_VALIDATION_DIR
+
 if [[ ${PROFILE} != *"scala-2.12"* ]]; then
-  ${CI_DIR}/license_check.sh $MVN_CLEAN_COMPILE_OUT $CI_DIR $(pwd) || exit $?
+  ${CI_DIR}/license_check.sh $MVN_CLEAN_COMPILE_OUT $CI_DIR $(pwd) $MVN_VALIDATION_DIR || exit $?
 fi
 
 exit $EXIT_CODE
